server/pyproject.toml
[project]
name = "voice-live-agent"
version = "0.1.0"
description = "Real-time voice agent with Azure Voice Live and ACS/Web streaming clients"
requires-python = ">=3.9"

dependencies = [
    "quart>=0.20.0",
    "websockets>=15.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.1.0",
    "quart-cors>=0.8.0",
    "azure-core>=1.34.0",
    "azure-communication-identity>=1.5.0",
    "aiohttp>=3.12.14",
    "azure-eventgrid>=4.22.0",
    "azure-identity>=1.23.0",
    "azure-communication-callautomation>=1.4.0",
    "azure-search-documents>=11.4.0",
    "openai[realtime]>=1.93.0",
    "h2>=4.3.0",
    "azure-storage-blob>=12.19.0",
    "pypdf>=3.17.0",
    "python-docx>=1.1.0",
    "PyJWT[crypto]>=2.8.0",
    "cryptography>=41.0.0"
]

[tool.uv]
# Define the virtual environment location for uv
# Note: project_environment is deprecated, use UV_PROJECT_ENVIRONMENT env var instead
